Textures can update.

+++ src/api/ogl/shader_data.cpp
@@ -30,7 +30,7 @@ shader_data_create(context_data *context, void *data)
     &context->shader_data_descs[cmd->shader_data_id]
   );
   assert(internal_desc);
-  
+
   // Check we have a name to search for.
   if(cmd->name == nullptr || strlen(cmd->name) == 0)
   {
@@ -53,7 +53,7 @@ shader_data_create(context_data *context, void *data)
         internal_desc->index    = uni_desc->index;
         internal_desc->type     = uni_desc->type;
         internal_desc->count    = uni_desc->count;
-        
+
         // Update In-out desc
         opShaderDataDesc *in_out_desc = cmd->desc;
         in_out_desc->count = internal_desc->count;
@@ -191,7 +191,7 @@ shader_data_bind(context_data *context, void *data)
     case(GL_SAMPLER_2D_SHADOW):
     #endif
     {
-      const size_t texture_id = static_cast<size_t>(cmd->data);
+      const uint32_t texture_id = static_cast<uint32_t>(cmd->data);
       const texture_desc *texture_desc(
         &context->texture_descs[texture_id]
       );
